Use an asteval interpreter
Make sure the temperature and pressure can be set in the symbol table.
Raise error if asteval detects invalid input.
Rework checks for OpenMC being available, allowing exceptions to be
raised if OpenMC materials aren't created successfully when OpenMC is
installed.
Implement a test to check that arbitrary code in the density_equation
results in an error.