Implement density as a MaterialProperty
Avoid impact on API (serialisation/deserialisation to/from JSON). This
requires some additional logic to ensure the density_unit is also
available.
Allow density to be populated as a MaterialProperty, or via a numeric
value in density, or via the density_equation.
Get the density_unit from the MaterialProperty.
Add a default_density property to get the density at the current
temperature and pressure of the material.
Create a Density property when calculating the density from OpenMC.